Fixed sane

+++ applications/sane.sh
@@ -49,9 +49,13 @@ whoami > /tmp/currentuser
 
 
 sudo tee rootscript.sh << "ENDOFROOTSCRIPT"
-rm -f /var/lock
-mkdir -pv /var/lock/
-touch /var/lock/sane
+if [ -L /var/lock ]; then
+	rm -f /var/lock
+fi
+if [ ! -d /var/lock ]; then
+	mkdir -pv /var/lock/
+	touch /var/lock/sane
+fi
 ENDOFROOTSCRIPT
 sudo chmod 755 rootscript.sh
 sudo bash -e ./rootscript.sh
